+++ src/commonMain/kotlin/com/jessecorbett/diskord/api/rest/client/internal/DefaultRestClient.kt
@@ -12,12 +12,11 @@ import io.ktor.client.request.*
 import io.ktor.client.request.forms.formData
 import io.ktor.client.request.forms.submitForm
 import io.ktor.client.request.forms.submitFormWithBinaryData
-import io.ktor.client.response.HttpResponse
+import io.ktor.client.statement.HttpResponse
 import io.ktor.http.ContentType
 import io.ktor.http.content.PartData
 import io.ktor.http.content.TextContent
-import kotlinx.coroutines.io.readUTF8Line
-import mu.KotlinLogging
+import io.ktor.utils.io.readUTF8Line
 
 private const val BOT_AUTH_PREFIX = "-> Authorization: Bot"
 
@@ -49,7 +48,9 @@ class DefaultRestClient(
     private val contentType = ContentType.parse("application/json")
 
     override suspend fun getRequest(url: String, headers: Map<String, String>): Response {
-        val result = client.get<HttpResponse>(baseUrl + url) { headers.forEach { header(it.key, it.value) } }
+        val result = client.get<HttpResponse>(baseUrl + url) {
+            headers.forEach { header(it.key, it.value) }
+        }
         return result.toResponse()
     }