+++ src/commonMain/kotlin/com/jessecorbett/diskord/util/Bytes.kt
@@ -1,7 +1,8 @@
 package com.jessecorbett.diskord.util
 
 import com.jessecorbett.diskord.api.rest.FileData
-import kotlinx.io.core.buildPacket
+import io.ktor.utils.io.core.buildPacket
+import io.ktor.utils.io.core.writeFully
 
 /**
  * Converts a [ByteArray] to a [FileData] object.